re-organize a bit
